+++ docs/src/examples/components/Avatar/Variations/AvatarExampleAlt.shorthand.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Avatar } from '@stardust-ui/react'
+
+const AvatarExampleAltShorthand = () => (
+  <Avatar src="/public/images/avatar/small/matt.jpg" alt="Profile picture of Matt" />
+)
+
+export default AvatarExampleAltShorthand